Resolved more review comments
